+++ market/src/webMain/kotlin/com/tneff/cyppie/market/HttpClientFactory.web.kt
@@ -1,0 +1,8 @@
+package com.tneff.cyppie.market
+
+import io.ktor.client.HttpClient
+import io.ktor.client.HttpClientConfig
+import io.ktor.client.engine.js.Js
+
+internal actual fun platformHttpClient(block: HttpClientConfig<*>.() -> Unit): HttpClient =
+    HttpClient(Js) { block() }